Html 759 - 3 (#226)
This is the next in a series of relatively large commits that have all contributed to the refactoring and reworking of the drug order tag in the 4.0.0 release of HFE.

Here are the features / changes this PR aims to implement:

Ensures that the tag respects the existing global property "drugOrder.requireDrug" (only allows leaving drug or drug non coded blank if this is false), and also adds additional validation to ensure concept is required and only one of drug and drugNonCoded can be supplied.

Adds the ability to configure any of the drug order properties that are Concepts (eg. concept, discontinueReason, orderReason, doseUnits, route, etc) via a single grouping concept (this can be either a question with answers or a set with members). Provide mechanisms to configure attributes within this mechanism to drive display labels (via message codes, concept name tags, and concept name types), css classes, and order of the resulting options rendered in within a dropdown widget.

Related to 2, modify the DropDownWidget and Option classes to support the ability to organize Options within groups and to apply css classes to both optgroup and option elements.

Changes the labels for the buttons used when editing an order within the original encounter it was placed in to "Edit" and "Delete" to reflect the implemented behavior which voids and recreates an Order in this circumstance rather than stop/discontinue. Adds in the ability to Delete an existing Discontinue Order in the encounter in which it was placed, in order to ensure appropriate data management is supported.

Related to 4, display a warning message letting users know that Edit and Delete will void previous orders when used within the current encounter.

Continued improvements and tweaks to the rendered drug order view, and improvements to usability.